refactor: extract trading into an opt-in plugin, lean api core
The api core (auth + wallet/vault + policy + proxy + webhooks) no longer
depends on the trading stack. The venue execution and trade-session routes,
their trade-specific auth middleware, and operator fund-recovery now live in
a separate package, registered onto the app via the StewardPlugin seam at the
composition root. The core does not import the plugin and the plugin does not
import the core: shared service singletons and auth middleware are injected
through the context the core hands register(app, ctx), so the dependency stays
one-directional with no cycle.

Behavior is unchanged: the same endpoints, auth gates, mount order, and
policy/audit invariants hold; a request that worked before works identically.
The core package now carries zero venue/trade-sessions dependencies, so an
auth-only install no longer ships the trading stack (smaller install, smaller
supply-chain and audit surface). This is phase 1 of a lean-core plus
opt-in-plugins architecture.
